moved routes into src directory
