+++ app.js
@@ -4,7 +4,7 @@ const http = require('http');
 const https = require('https');
 const methodOverride = require('method-override');
 const morgan = require('morgan');
-const routes = require('./routes');
+const routes = require('./src/routes');
 const port = global.port = process.env.PORT || 3100;
 const env = process.env.NODE_ENV || 'development';
